Move `IVertexBuffer` and add generic variant
