+++ osu.Framework/Graphics/Batches/DepthWrappingVertex.cs
@@ -5,7 +5,7 @@
 
 using System;
 using System.Runtime.InteropServices;
-using osu.Framework.Graphics.OpenGL.Vertices;
+using osu.Framework.Graphics.Batches;
 using osuTK.Graphics.ES30;
 
 namespace osu.Framework.Graphics.Batches
