+++ osu.Framework/Graphics/Batches/TexturedVertex2D.cs
@@ -5,7 +5,7 @@
 
 using System;
 using System.Runtime.InteropServices;
-using osu.Framework.Graphics.OpenGL.Vertices;
+using osu.Framework.Graphics.Batches;
 using osuTK;
 using osuTK.Graphics;
 using osuTK.Graphics.ES30;
