+++ osu.Framework/Graphics/Rendering/IVertexBatch.cs
@@ -2,7 +2,7 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using System;
-using osu.Framework.Graphics.OpenGL.Vertices;
+using osu.Framework.Graphics.Batches;
 
 namespace osu.Framework.Graphics.Rendering
 {
